use custom visit with delay and reorganize
