+++ platform_umbrella/apps/verify/test/support/test_case.ex
@@ -2,28 +2,32 @@ defmodule Verify.TestCase do
   @moduledoc false
   use ExUnit.CaseTemplate
 
-  alias CommonCore.Batteries.Catalog
-  alias Verify.BatteryInstallWorker
-
-  require Logger
-
   using options do
     install_spec = Keyword.get(options, :install_spec, :int_test)
     batteries = Keyword.get(options, :batteries, [])
     images = Keyword.get(options, :images, [])
 
     quote do
-      use Wallaby.DSL
+      unquote(__prelude(batteries, images))
+      unquote(__setup_all(batteries, images, install_spec))
+      unquote(__setup())
+    end
+  end
 
-      import Verify.TestCase,
-        only: [
-          start_session: 0,
-          install_batteries: 2,
-          uninstall_batteries: 2
-        ]
-
+  def __prelude(batteries, images) do
+    quote do
+      # copied from `use Wallaby.DSL`
+      # we want to override visit/2
+      import Kernel, except: [tap: 2]
       import Verify.TestCase.Helpers
+      import Verify.TestCase.Util
+      import Wallaby.Browser, except: [visit: 2]
 
+      alias Wallaby.Browser
+      alias Wallaby.Element
+      alias Wallaby.Query
+
+      # Kernel.tap/2 was introduced in 1.12 and conflicts with Browser.tap/2
       require Logger
 
       ExUnit.Case.register_module_attribute(__MODULE__, :batteries, accumulate: true)
@@ -34,6 +38,12 @@ defmodule Verify.TestCase do
 
       @moduletag :cluster_test
 
+      @args ["--class=bi-int-test"]
+    end
+  end
+
+  def __setup_all(batteries, images, install_spec) do
+    quote do
       setup_all do
         batteries = unquote(batteries)
         images = unquote(images)
@@ -47,7 +57,7 @@ defmodule Verify.TestCase do
           })
 
         # kick off image pull as early as possible
-        unquote(__MODULE__).prepull_images(image_pid, images)
+        prepull_images(image_pid, images)
 
         Logger.debug("Starting Kind for spec: #{unquote(install_spec)}")
         tmp_dir = get_tmp_dir(__MODULE__)
@@ -94,101 +104,17 @@ defmodule Verify.TestCase do
            tmp_dir: tmp_dir
          ]}
       end
+    end
+  end
 
+  def __setup do
+    quote do
       setup do
         {:ok, session} = start_session()
-        on_exit(fn -> Wallaby.end_session(session) end)
 
+        on_exit(fn -> Wallaby.end_session(session) end)
         {:ok, [session: session]}
       end
     end
-  end
-
-  def prepull_images(_pid, []), do: :ok
-
-  def prepull_images(pid, images) do
-    Enum.each(images, &Verify.ImagePullWorker.pull_image(pid, &1))
-  end
-
-  def install_batteries(worker_pid, batteries \\ [])
-
-  def install_batteries(worker_pid, battery) when is_atom(battery), do: install_battery(worker_pid, {battery, %{}})
-
-  def install_batteries(_worker_pid, []), do: :ok
-
-  def install_batteries(worker_pid, batteries) do
-    Enum.map(batteries, &install_battery(worker_pid, &1))
-  end
-
-  defp install_battery(pid, type) when is_atom(type), do: install_battery(pid, {type, %{}})
-
-  defp install_battery(pid, {type, config}) do
-    case Catalog.get(type) do
-      nil ->
-        raise "Couldn't find battery: #{inspect(type)}"
-
-      battery ->
-        BatteryInstallWorker.install_battery(pid, battery, config)
-    end
-  end
-
-  def uninstall_batteries(worker_pid, battery) when is_atom(battery), do: uninstall_batteries(worker_pid, [battery])
-
-  def uninstall_batteries(_worker_pid, []), do: :ok
-
-  def uninstall_batteries(worker_pid, batteries) do
-    Enum.map(batteries, fn type ->
-      case Catalog.get(type) do
-        nil ->
-          raise "Couldn't find battery: #{inspect(type)}"
-
-        battery ->
-          BatteryInstallWorker.uninstall_battery(worker_pid, battery)
-      end
-    end)
-  end
-
-  @spec start_session() :: {:ok, Wallaby.Session.t()} | {:error, Wallaby.reason()}
-  def start_session do
-    Wallaby.start_session(
-      max_wait_time: 60_000,
-      capabilities: %{
-        headless: true,
-        javascriptEnabled: true,
-        loadImages: true,
-        chromeOptions: %{
-          args: [
-            # Lets act like the world is run on macbooks that
-            # all of sillion valley uses
-            #
-            # Fix this at some point
-            "window-size=1920,1080",
-            # We don't want to see the browser
-            "--headless",
-            "--fullscreen",
-            # Incognito mode means no caching for real
-            # Unfortunately, chrome doesn't allow http requests at all incognito
-            # "--incognito",
-            # Seems to be better for stability
-            "--no-sandbox",
-            # Yeah this will run in CI
-            "--disable-gpu",
-            # Disable dev shm usage
-            # This is needed for CI environments like GitHub Actions
-            # where /dev/shm is super small
-            #
-            # See
-            # https://github.com/elixir-wallaby/wallaby/issues/468#issuecomment-1113520767
-            "--disable-dev-shm-usage",
-            # Please google go away
-            "--disable-extensions",
-            "--disable-login-animations",
-            "--no-default-browser-check",
-            "--no-first-run",
-            "--ignore-certificate-errors"
-          ]
-        }
-      }
-    )
   end
 end
